Added DAO criteria API method for checking if similarity already exists

+++ sdr-core/src/main/java/ba/com/zira/sdr/core/validation/SongSimilarityRequestValidation.java
@@ -23,7 +23,7 @@ public class SongSimilarityRequestValidation {
     }
 
     private ValidationError validation(Long songA, Long songB) {
-        if (songA == songB) {
+        if (songA.longValue() == songB.longValue()) {
             return ValidationError.of("IDENTICAL_SONGS", "Cannot set song similarity to itself!");
         }
